Added: New GNU AS ARM Assembly lexer
First working lexer file that can be used to parse ARMv7 Assembly
scripts, written using the simple GNU AS syntax (not the AT&T one).
This MIGHT be usable to highlight assembly scripts written for
Keil's ARM Assembler, but since I do not own it, I cannot attest.

Anyway, GNU AS for ARM is different from GNU AS for Intel, and still,
it might need its own lexer. This lexer could then *include* other
lexer class files in order to specialise it to a specific Assembly
language and syntax. (class GNU_AS < Lexer; include GNU_AS_ARM; end)

Still, it works fine for simple GNU AS ARM scripts, and that's fine
for me.

Signed-off-by: Myy <myy@miouyouyou.fr>