Rearchitectured the GNU AS ARM Lexer
This rearchitecture job should make it more easy to add new GAS Lexers
for other syntaxes, and architecture mnemonics.

This also added in complexity, a part of it is due to the fact that I
do not get how the Lexers work, and I do not intend to pass a full day
understanding the code.

I'll test this parser a few days. If it works, I'll add a few examples,
some tests and generate a pull request.

Signed-off-by: Myy <myy@miouyouyou.fr>