Added an HTML formatter. Corrected a bug with GNU AS lexer.
Corrected a bug that caused the lexer to not parse the instructions
of the last line if it did not contain a carriage return.

Added an HTML formatter, HXML, that outputs tokens as tags, with
attributes specified as tag attributes. This is clearly a personal
preference and I'll avoid pushing this change with the GNU AS ARM
lexer.
That said, the improved semantic really help make things clearer,
compared to the 'span' fest generated by traditionals HTML formatters.

Also modified the gemspec so that it includes the files required by the
GNU AS ARM lexer.

Signed-off-by: Myy <myy@miouyouyou.fr>